+++ Tests/Unit/Stubs/Microsoft365.psm1
@@ -11637,6 +11637,31 @@ function Get-CsTenantDialPlan
         $AsJob
     )
 }
+function Get-CsTenantFederationConfiguration
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $LocalStore,
+
+        [Parameter()]
+        [System.Object]
+        $Identity,
+
+        [Parameter()]
+        [System.Object]
+        $Tenant,
+
+        [Parameter()]
+        [System.Object]
+        $Filter,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $AsJob
+    )
+}
 function Grant-CsTeamsUpgradePolicy
 {
     [CmdletBinding()]
@@ -14056,6 +14081,43 @@ function Set-CsTenantDialPlan
         [Parameter()]
         [System.Management.Automation.SwitchParameter]
         $AsJob
+    )
+}
+function Set-CsTenantFederationConfiguration
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.Object]
+        $AllowFederatedUsers,
+
+        [Parameter()]
+        [System.Object]
+        $AllowPublicUsers,
+
+        [Parameter()]
+        [System.Object]
+        $Identity,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $Confirm,
+
+        [Parameter()]
+        [System.Object]
+        $AllowTeamsConsumer,
+
+        [Parameter()]
+        [System.Object]
+        $AllowTeamsConsumerInbound,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $Force,
+
+        [Parameter()]
+        [System.Object]
+        $Tenant
     )
 }
 #endregion