+++ CHANGELOG.md
@@ -15,8 +15,8 @@
   * Revamped to use Microsoft Graph in the Set;
 * TeamsChannel
   * Fix to retrieve the team name without URL encoding.
-* M365DSCReport
-  * Adds JSON as an export option.
+* TeamsFederationConfiguration
+  * New resource for Teams Federation Configuration.
 * TeamsTeam
   * Fixed issue where teams were not created when no owners were specified.
     If credentials are used, then the user will be used as owner.
@@ -24,7 +24,8 @@
   * Updated DSCParser to version 1.3.0.3.
 * MISC
   * Added a function to uninstall all outdated dependencies
-    and older versions of Microsoft365DSC
+    and older versions of Microsoft365DSC;
+  * M365DSCReport: Adds JSON as an export option.
 
 # 1.22.216.1